requirements.txt
numpy
umap-learn
hdbscan
plotly